+++ tests/test_azure_store.py
@@ -55,6 +55,7 @@ def _delete_container(conn_string, container):
             # ignore the ContainerNotFound error:
             if ex.error_code != "ContainerNotFound":
                 raise
+        s.close()
     except ImportError:
         # for azure-storage-blob<12
         from azure.storage.blob import BlockBlobService
@@ -68,9 +69,10 @@ class TestAzureStorage(BasicStore, OpenSeekTellStore):
     def store(self):
         container = str(uuid())
         conn_string = get_azure_conn_string()
-        yield AzureBlockBlobStore(
+        with AzureBlockBlobStore(
             conn_string=conn_string, container=container, public=False
-        )
+        ) as store:
+            yield store
         _delete_container(conn_string, container)
 
 
@@ -87,10 +89,54 @@ class TestExtendedKeysAzureStorage(TestAzureStorage, ExtendedKeyspaceTests):
         class ExtendedKeysStore(ExtendedKeyspaceMixin, AzureBlockBlobStore):
             pass
 
-        yield ExtendedKeysStore(
+        with ExtendedKeysStore(
             conn_string=conn_string, container=container, public=False
-        )
+        ) as store:
+            yield store
         _delete_container(conn_string, container)
+
+
+@pytest.mark.filterwarnings("error")
+def test_azure_dangling_port_enter_exit():
+    container = str(uuid())
+    conn_string = get_azure_conn_string()
+    with AzureBlockBlobStore(conn_string=conn_string, container=container) as store:
+        if not hasattr(store, "blob_container_client"):
+            # This test only runs for azurestore_new
+            return
+        store.blob_container_client  # type: ignore
+
+
+@pytest.mark.filterwarnings("error")
+def test_azure_dangling_port_explicit_close():
+    container = str(uuid())
+    conn_string = get_azure_conn_string()
+    store = AzureBlockBlobStore(conn_string=conn_string, container=container)
+    if not hasattr(store, "blob_container_client"):
+        # This test only runs for azurestore_new
+        return
+    store.blob_container_client  # type: ignore
+    store.close()
+
+
+@pytest.mark.filterwarnings("error")
+def test_azure_dangling_port_explicit_close_multi():
+    container = str(uuid())
+    conn_string = get_azure_conn_string()
+    store = AzureBlockBlobStore(conn_string=conn_string, container=container)
+    if not hasattr(store, "blob_container_client"):
+        # This test only runs for azurestore_new
+        return
+    store.blob_container_client  # type: ignore
+    # We check that multiclose and reuse do not do weird things
+    store.close()
+    store.close()
+    store.blob_container_client  # type: ignore
+    store.blob_container_client  # type: ignore
+    store.blob_container_client  # type: ignore
+    store.close()
+    store.close()
+    store.close()
 
 
 def test_azure_setgetstate():
@@ -100,10 +146,12 @@ def test_azure_setgetstate():
     store.put("key1", b"value1")
 
     buf = pickle.dumps(store, protocol=2)
+    store.close()
     store = pickle.loads(buf)
 
     assert store.get("key1") == b"value1"
     _delete_container(conn_string, container)
+    store.close()
 
 
 def test_azure_store_attributes():
@@ -144,6 +192,7 @@ def test_azure_special_args():
         cfg = abbs.blob_container_client._config  # type: ignore
         assert cfg.max_single_put_size == MSP
         assert cfg.max_block_size == MBS
+    abbs.close()
 
 
 class TestAzureExceptionHandling:
@@ -156,6 +205,7 @@ class TestAzureExceptionHandling:
         with pytest.raises(IOError) as exc:
             store.keys()
         assert "The specified container does not exist." in str(exc.value)
+        store.close()
 
     def test_wrong_endpoint(self):
         container = str(uuid())
@@ -176,6 +226,7 @@ class TestAzureExceptionHandling:
         with pytest.raises(IOError) as exc:
             store.put("key", b"data")
         assert "connect" in str(exc.value)
+        store.close()
 
     def test_wrong_credentials(self):
         container = str(uuid())
@@ -198,6 +249,7 @@ class TestAzureExceptionHandling:
         with pytest.raises(IOError) as exc:
             store.put("key", b"data")
         assert "Incorrect padding" in str(exc.value)
+        store.close()
 
 
 class TestChecksum:
@@ -210,12 +262,13 @@ class TestChecksum:
         container = str(uuid())
         conn_string = get_azure_conn_string()
 
-        yield AzureBlockBlobStore(
+        with AzureBlockBlobStore(
             conn_string=conn_string,
             container=container,
             public=False,
             checksum=True,
-        )
+        ) as store:
+            yield store
         _delete_container(conn_string, container)
 
     def _checksum(self, store):
@@ -241,7 +294,8 @@ class TestChecksum:
     def test_checksum_put_file(self, store, tmpdir):
         file_ = tmpdir.join("my_file")
         file_.write(self.CONTENT)
-        store.put_file(self.KEY, file_.open("rb"))
+        with file_.open("rb") as infile:
+            store.put_file(self.KEY, infile)
         assert self._checksum(store) == self.EXPECTED_CHECKSUM
         assert store.get(self.KEY) == self.CONTENT