+++ docs/changes.rst
@@ -1,6 +1,13 @@
 Changelog
 *********
 
+1.5.0
+=====
+
+* Added concept for closable stores.
+  * Stores and Decorators can now be opened using with KeyValueStore as store
+  * Implemented this functionality for baseclasses and the AzureBlockBlobStore
+
 1.4.4
 =====